services/user_management/requirement.txt
asgiref==3.8.1
async-timeout==4.0.3
certifi==2024.8.30
charset-normalizer==3.4.0
Django==4.2.16
django-environ==0.11.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
idna==3.10
oauthlib==3.2.2
pillow==10.4.0
PyJWT==2.9.0
redis==5.2.0
requests==2.32.3
requests-oauthlib==2.0.0
sqlparse==0.5.1
typing_extensions==4.12.2
urllib3==2.2.3